requirements.txt
fastapi
python-dotenv
openai
aiohttp
requests
twilio
python-multipart

